My first game
This started out as an if,else test, but ended up i a tiny guessing game. It is not much, but I thought it was fun and it was all done in just an hour, which was the time it took my wifi to cook dinner. I only put this up for personal reasons since it is my first actual tiny game developed without any tutorials or help. A memorable little piece of code which can be fun to look back after some time.